+++ livestatus/src/Logfile.h
@@ -54,7 +54,7 @@ private:
     time_t _since;     // time of first entry
     bool _watch;       // true only for current logfile
     fpos_t _read_pos;  // read until this position
-    uint32_t _lineno;  // read until this line
+    size_t _lineno;    // read until this line
 
     logfile_entries_t _entries;
 #ifdef CMC
@@ -92,8 +92,8 @@ public:
 private:
     void loadRange(FILE *file, unsigned missing_types, LogCache *, time_t since,
                    time_t until, unsigned logclasses);
-    bool processLogLine(uint32_t lineno, std::string line, unsigned logclasses);
-    uint64_t makeKey(time_t, unsigned);
+    bool processLogLine(size_t lineno, std::string line, unsigned logclasses);
+    uint64_t makeKey(time_t, size_t);
     Logger *logger() const;
 };
